87377_test 통과

+++ src/test/java/com/ll/programmers/Lv2/P87377/Tests.java
@@ -13,7 +13,7 @@ public class Tests {
                         new int[]{2, -1, 0}
                 )
         ).isEqualTo(
-                new long[]{0, 0}
+                new int[]{0, 0}
         );
     }
